pyproject.toml
[tool.poetry]
name = "powerline-k8s"
version = "1.5.2"
description = "A custom Powerline segment for displaying the current Kubernetes context and namespace"
authors = ["Jordan Duabe <me@jduabe.dev>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/j4ckofalltrades/powerline-k8s"
repository = "https://github.com/j4ckofalltrades/powerline-k8s"
keywords = ["powerline", "kubernetes"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/j4ckofalltrades/powerline-k8s/issues"

[tool.poetry.dependencies]
python = "^3.9"
kubernetes = "^29.0.0"
powerline-status = "^2.7"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^12.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
